+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.0.0
 
 - Use own twitch name setting (setting server tab)
+- Poll once on game start, stop if not live.  Click to poll.
 
 ## 1.1.0